+++ scripts/get_native_libraries.py
@@ -1,5 +1,6 @@
 # -*- coding: utf-8 -*-
 import urllib.request
+import os
 import shutil
 import ssl
 import zipfile
@@ -7,7 +8,7 @@ import zipfile
 from os import path, remove, system
 
 
-def download_native_libraries(name: str, version: str, destination: str):
+def download_native_libraries(name: str, version: str, destination: str) -> bool:
     ssl._create_default_https_context = ssl._create_unverified_context
     jni_libs = 'android/src/main/jniLibs'
 
@@ -57,11 +58,15 @@ def download_native_libraries(name: str, version: str, destination: str):
                 remove('all.zip')
         except Exception as e:
             print(f'Cannot get {name} {version} ({e})')
-            download_native_libraries(name, 'last_build', destination)
+            return False
+    return True
 
 
 if __name__ == '__main__':
-    download_native_libraries('findex', 'v2.0.1', 'resources')
-    download_native_libraries('findex', 'last_build', 'resources')
-    download_native_libraries('cover_crypt', 'v8.0.2', 'resources')
-    download_native_libraries('cover_crypt', 'last_build', 'resources')
+    ret = download_native_libraries('findex', 'v2.0.1', 'resources')
+    if ret is False and os.getenv('GITHUB_ACTIONS'):
+        download_native_libraries('findex', 'last_build', 'resources')
+
+    ret = download_native_libraries('cover_crypt', 'v10.0.0', 'resources')
+    if ret is False and os.getenv('GITHUB_ACTIONS'):
+        download_native_libraries('cover_crypt', 'last_build', 'resources')